make PromFed and HPO charts doNotRelease:true
We won't need these as set to true anymore.
Initially I did when we needed to keep them here because they depend on each other and need the chart build scripts.
However, we opted to merge into a single chart that PromFed releases and publish 1:1 like we do with BRO, so we can ship to rancher/charts via release tags.